Registration Form - full Code
Create a Input component that is dynamic and highly customizable. Customizable in the sense that it can be used in place of different input types.
Built with HTML, CSS, React JS .
The following components must be created
Input
InputComponent

The input component provides the props to each of the Input component or rather it renders the Input Components in the first place.
Here Each Input along with its label and the error message is an individual Input Component. This input component can be used for the following types:
text
email
password.
The input validation should not be triggered unless each of its input is focused.